// src/components/Table/types.ts
export type Row = Record<string, string> & {
	_id?: string | number;
	id: string | number;
};

// Manage Columns
export type TComponentType =
	| 'link'
	| 'date'
	| 'dateTime'
	| 'duration'
	| 'status'
	| 'time';
// ...
export interface TPaginationProps {
	total: number;
	page: number;
	limit: number;
}

export type ITable = TableProps & {
	columns: IColumns;
	rows: Row[];
	getters: {
		[key: string]: number | string;
	};
	description?: string;
	onSaveColumn: () => void;
	rowActions?: () => React.ReactNode;
	showRowDetails?: boolean;
	rowDetails?: () => React.ReactNode;
	fetchData: (
		_pagination: TPaginationProps,
		_sortOrder: any,
		_search: string,
		_appliedFilters: any
	) => void;
	rowsPerPage?: number | undefined;
	defaultPage?: number | undefined;
	selectable?: boolean;
	showManageColumns?: boolean;
	renderRow?: () => React.ReactNode;
	actions?: JSX.Element | null | undefined;
	tableName?: string;
	tableView?: string;
	emptyComponent?: JSX.Element | null | undefined;
	onRowClick?: () => void;
	refetchData?: any[];
	showSearch?: boolean;
	showFilters?: boolean;
	selectedRowPaths?: string[];
	setSelectedRowPaths?: (Args: any) => void;
	filters: {
		[key: string]: any;
	};
	showBorder?: boolean;
	showFooter?: boolean;
	columnWidth?: string | number;
	singleSort?: boolean;
	renderAttachment?: JSX.Element;
	linkComponent?: JSX.Element;
	rowsPerPageOptions?: number[];
	manageColumnButtonProps?: Object;
	onPageChange?: (page: number) => {};
	onRowsPerPageChange?: (pageSize: number) => {};
	showTableMeta?: boolean;
};
// ...
export interface IColumn {
	type?: TComponentType;
	key: string;
	options?: Record<string, string>;
	field?: string;
	isVisible?: boolean;
	label: string;
	sortable?: boolean;
	colSpan?: number | string;
	render: () => React.ReactNode;
	isMandatory?: boolean;
}

export type IColumns = IColumn[];